+++ 5Things/src/com/packages/things/java/Shirt.java
@@ -11,9 +11,9 @@ public class Shirt {
 	public final String companyName="Biba";
 	public ShirtBrand brand;
 	public String address;
-	public double price =500;
+	public Integer price =500;
 	public long contacuNo;
-	public void setShirt(ShirtBrand brand, String address, double price) {
+	public void setShirt(ShirtBrand brand, String address, Integer price) {
 		this.brand = brand;
 		this.address = address;
 		this.price = price;
