+++ Interface/src/com/xworkz/interfaces/rules/HospitalManagement.java
@@ -2,7 +2,7 @@ package com.xworkz.interfaces.rules;
 
 public interface HospitalManagement {
 	
-	public double entryPrice();
+	public Integer entryPrice();
 	public boolean followingRules();
