+++ Interface/src/com/xworkz/interfaces/rules/KarnatakaRTO.java
@@ -2,7 +2,7 @@ package com.xworkz.interfaces.rules;
 
 public interface KarnatakaRTO extends TransportRule {
 	
-	public double vat();
+	public Integer vat();
 	
 
 }
