+++ Interface/src/com/xworkz/interfaces/rules/TransportRule.java
@@ -2,7 +2,7 @@ package com.xworkz.interfaces.rules;
 
 public interface TransportRule {
 	public boolean paidTax();
-	double gst();
+	Integer gst();
 	
 
 }
