+++ constructor/src/com/xworkz/things/Fish.java
@@ -4,9 +4,9 @@ public class Fish {
 	
 	public String name;
 	public  String type;
-	public double price;
-	public  double weight;
-	public double height;
+	public Integer price;
+	public  Integer weight;
+	public Integer height;
 	
 	public Fish(String name ,String type)
 	{
@@ -14,33 +14,33 @@ public class Fish {
 		this.type=type;
 		
 	}
-	public Fish(String name ,double  price)
+	public Fish(String name ,Integer  price)
 	{
 		this.name=name;
 		this.price=price;
 		
 	}
-	public Fish(double price ,double weight)
+	public Fish(Integer price ,Integer weight)
 	{
 		this.price=price;
 		this.weight=weight;
 		
 	}
-	public Fish(double  weight ,double height ,String name)
+	public Fish(Integer  weight ,Integer height ,String name)
 	{
 		this.weight=weight;
 		this.height=height;
 		this.name=name;
 		
 	}
-	public Fish(String name ,double price,String type)
+	public Fish(String name ,Integer price,String type)
 	{
 		this.name=name;
 		this.price=price;
 		this.type=type;
 		
 	}
-	public Fish(double price ,String name, String type)
+	public Fish(Integer price ,String name, String type)
 	{
 		this.price=price;
 		this.name=name;
