+++ constructor/src/com/xworkz/things/Tractor.java
@@ -4,7 +4,7 @@ public class Tractor {
 	
 	public String name;
 	public String brand;
-	public double price;
+	public Integer price;
 	public String color;
 	public int hp;
 	public String model;
@@ -17,30 +17,30 @@ public class Tractor {
 		this.name=name;
 		
 	}
-	public Tractor( double price,String brand)
+	public Tractor( Integer price,String brand)
 	{
 		this("SONALIKA TRACTOR");
 		this.price=price;
 		this.brand=brand;
 	}
-	public Tractor( String color,String brand,double price)
+	public Tractor( String color,String brand,Integer price)
 	{
 		this(100.0,brand);
 		this.color=color;
 	}
-	public Tractor( double price ,String color,String brand,int hp)
+	public Tractor( Integer price ,String color,String brand,int hp)
 	{
 		this(color,"International",price);
 		this.hp=hp;
 	}
-	public Tractor( String model,int hp,double price,String brand)
+	public Tractor( String model,int hp,Integer price,String brand)
 	{
 		this(price,"RED",brand,41);
 		this.hp=hp;
 		
 		this.model=model;
 	}
-	public Tractor( String name ,int hp,double price,String brand,String color,String model)
+	public Tractor( String name ,int hp,Integer price,String brand,String color,String model)
 	{
 		this(model,hp,price,brand);
