+++ constructor/src/com/xworkz/things/boot/CrecketRunner.java
@@ -9,7 +9,7 @@ public class CrecketRunner {
 		 String alroungPylayar []= {"Hardik pandya","aksharPatel","Kapil Dev","jadeja"};
 		 String bolers[]= {"Bhuvanesh kumar","mahamd shami","Shardul takur","Chahal"};
 		String weketkeepers[]= {"MS doni","Dinesh karthik","Rishabpanth"};
-		 double salary[]= {1000000,230000,450000,40000,500000};
+		 Integer salary[]= {1000000,230000,450000,40000,500000};
 		 
 		 Crecket crecket=new Crecket("Banglore", playersName, "Chinnaswami Stadioum", alroungPylayar, 3000, "Green", bolers, 1500, weketkeepers, true, salary);
 		 crecket.display();
